% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDAS_functions.R
\name{anova_test}
\alias{anova_test}
\title{Run Model Test and Return p-value and Statistic}
\usage{
anova_test(data, formula0, formula1, model_fit_function)
}
\arguments{
\item{data}{A data frame containing the data for the models.}

\item{formula0}{A character string representing the formula for the null model.}

\item{formula1}{A character string representing the formula for the alternative model.}

\item{model_fit_function}{A character string specifying the model fitting function to use ("lm" or "lmer").}
}
\value{
A numeric vector of length 2, where the first element is the p-value and the second element is the test statistic.
}
\description{
Fits two models using either \code{lm} or \code{lmer} (if a random effect is provided)
and performs an ANOVA to compare them. Returns the p-value and a test statistic.
}
\examples{
\dontrun{
# Using linear models:
run_test(data = my_data, "Y ~ x", "Y ~ x + z", model_fit_function = "lm")
}
}
\keyword{internal}
