% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeway_functions.R
\name{check_factor_name3}
\alias{check_factor_name3}
\title{Check the iDAS_3F input factors' name}
\usage{
check_factor_name3(f1name, f2name, f3name, randomname, f1, f2, f3, random)
}
\arguments{
\item{f1name}{a string of factor 1's name}

\item{f2name}{a string of factor 2's name}

\item{f3name}{a string of factor 3's name}

\item{randomname}{a string of random effect term's name}

\item{f1}{A vector of factor 1 variables.}

\item{f2}{A vector of factor 2 variables.}

\item{f3}{A vector of factor 3 variables.}

\item{random}{A vector of random effect term variables.}
}
\value{
A list of each factor's name and values
}
\description{
This function will be no longer be used.
}
\examples{
 #factor_tmp=check_factor_name3(f1name,f2name,f2name,randomname,f1,f2,f3,random)
}
\keyword{internal}
