% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_names}
\alias{collapse_names}
\title{Collapse Names}
\usage{
collapse_names(name_vec)
}
\arguments{
\item{name_vec}{A character vector of names.}
}
\value{
A single string with names separated by "+" if there is more than one name,
otherwise returns the single name.
}
\description{
Collapses a vector of names into a single string, concatenated with a plus sign.
}
\keyword{internal}
