% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_default_name}
\alias{set_default_name}
\title{Set Default Name}
\usage{
set_default_name(name, default)
}
\arguments{
\item{name}{A character value representing the current name.}

\item{default}{A character value representing the default name to use if \code{name} is NULL.}
}
\value{
The original \code{name} if not NULL, otherwise \code{default}.
}
\description{
Returns the default name if the provided name is NULL.
}
\keyword{internal}
