% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_functions.R
\name{check_factor_name}
\alias{check_factor_name}
\title{Check the iDAS_2F input factors' name}
\usage{
check_factor_name(f1name, f2name, randomname, f1, f2, random)
}
\arguments{
\item{f1name}{a string of factor 1's name}

\item{f2name}{a string of factor 2's name}

\item{randomname}{a string of random effect term's name}

\item{f1}{A vector of factor 1 variables.}

\item{f2}{A vector of factor 2 variables.}

\item{random}{A vector of random effect term variables.}
}
\value{
A list of each factor's name and values
}
\description{
Check the iDAS_2F input factors' name
}
\examples{
 #factor_tmp=check_factor_name(f1name,f2name,randomname,f1,f2,random)
}
