% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_functions.R
\name{check_factor_name2}
\alias{check_factor_name2}
\title{Check the iDAS_2F input factors' name}
\usage{
check_factor_name2(
  factor1_name,
  factor2_name,
  random_effect_name,
  factor1,
  factor2,
  random_effect
)
}
\arguments{
\item{factor1_name}{A string for the first factor variable's name.}

\item{factor2_name}{A string for the second factor variable's name.}

\item{random_effect_name}{A string for the random effect term's name.}

\item{factor1}{A vector of the first factor variable.}

\item{factor2}{A vector of the second factor variable.}

\item{random_effect}{A vector of the random effect term variables.}
}
\value{
A list of each factor's name and values.
}
\description{
This function will be no longer be used.
}
\examples{
# formatted_factors = check_factor_name(factor1_name, factor2_name, random_effect_name, factor1, factor2, random_effect)
}
\keyword{internal}
