% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_functions.R
\name{iDAS_2F}
\alias{iDAS_2F}
\title{Interpretable differential abundance analysis (two-way analysis)}
\usage{
iDAS_2F(
  Z,
  f1,
  f2,
  random = NULL,
  test_func = "lm",
  Sig_cutoff = 0.02,
  Sig = 0.05,
  Int = 0.01,
  F1 = 0.01,
  F2 = 0.01,
  adj_method = "BH",
  f1name = NULL,
  f2name = NULL,
  randomname = NULL
)
}
\arguments{
\item{Z}{A matrix/dataframe of omics or gene expression data, row as sample.}

\item{f1}{A vector of factor 1 variables.}

\item{f2}{A vector of factor 2 variables.}

\item{random}{A vector of random effect term of ANOVA analysis,
by default is NULL, which means the model doesn't include random effect term.}

\item{test_func}{Testing function used, either stats::lm or lme4::lmer. By default is "lm".}

\item{Sig_cutoff}{No effect test significance level is defined by a fraction value
to indicate when ordering the p-values
and defining the top X\% as the significance level. If both Sig_cutoff and Sig
are set, the algorithm will, by default, use Sig_cutoff to determine
the significance level.}

\item{Sig}{No effect test significance level is defined directly
by a fraction value, by default is 0.05}

\item{Int}{Interaction effect test significance level,by default is 0.01}

\item{F1}{F1 effect test significance level, by default is 0.01}

\item{F2}{F2 effect test significance level, by default is 0.01}

\item{adj_method}{Pvalue adjust method. See p.adjust. By default is "BH".}

\item{f1name}{The column name of factor 1, by default is F1.}

\item{f2name}{The column name of factor 2, by default is F2.}

\item{randomname}{The column name of random effect term, by default is Random.}
}
\value{
A list of hypothesis test outcome, P_mat is the pvalue matrix of all tests,
S_mat is the statistics matrix of all test, cls_df is the Classification data frame of all tests.
}
\description{
Interpretable differential abundance analysis (two-way analysis)
}
\examples{
# res=iDAS_2F(Z= X,
#f1=pcelltype,f2=pcell_stats,random=NULL,test_func="lm",
#Sig_cutoff=0.02, Sig = 0.1,Int= 0.01, F1 = 0.01,F2 = 0.01,
#adj_method="BH",f1name=NULL,f2name=NULL,randomname=NULL)
}
