% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_input}
\alias{process_input}
\title{Process Input}
\usage{
process_input(x, name)
}
\arguments{
\item{x}{A factor or vector that should be converted.}

\item{name}{A character value to assign as the column name.}
}
\value{
A data frame with one column named \code{name} containing \code{x}.
}
\description{
Converts a factor or vector to a data frame and assigns a specified column name.
}
\keyword{internal}
